# ZBP-polish-api-2_1_2-swagger.json
{
  "swagger" : "2.0",
  "info" : {
    "version" : "2_1_2",
    "title" : "Polish API",
    "description" : "Interface specification for services provided by third parties based on access to payment accounts. Prepared by the Polish Bank Association and its affiliates"
  },
  "host" : "apiHost",
  "basePath" : "/",
  "schemes" : [ "https" ],
  "consumes" : [ "application/json" ],
  "produces" : [ "application/json" ],
...
  "parameters" : {
...
    "acceptEncodingParam" : {
      "name" : "Accept-Encoding",
      "in" : "header",
      "required" : true,
      "type" : "string",
      "enum" : [ "gzip", "deflate" ],
      "description" : "Gzip, deflate"
    },
    "acceptLanguageParam" : {
      "name" : "Accept-Language",
      "in" : "header",
      "required" : true,
      "type" : "string",
      "description" : "Prefered language of response"
    },
    "acceptCharsetParam" : {
      "name" : "Accept-Charset",
      "in" : "header",
      "required" : true,
      "type" : "string",
      "enum" : [ "utf-8" ],
      "description" : "UTF-8"
    },
    "xjwsSignatureParam" : {
      "name" : "X-JWS-SIGNATURE",
      "in" : "header",
      "required" : true,
      "type" : "string",
      "description" : "Detached JWS signature of the body of the payload"
    },
    "xRequestIdParam" : {
      "name" : "X-REQUEST-ID",
      "in" : "header",
      "required" : true,
      "type" : "string",
      "description" : "Identyfikator żądania w formacie UUID (Wariant 1, Wersja 1), zgodnym ze standardem RFC 4122, nadawany przez TPP. Wartość musi być zgodna z parametrem requestId przekazywanym w ciele każdego żądania. / Request identifier using UUID format (Variant 1, Version 1), described in RFC 4122 standard, set by TPP. Value of this header must be the same as for the requestId param passed inside request payload."
    }
  },
  "paths" : {
...
    "/v2_1_2.1/confirmation/v2_1_2.1/getConfirmationOfFunds" : {
      "x-swagger-router-controller" : "caf",
      "post" : {
        "summary" : "Confirmation of the availability of funds",
        "description" : "Confirming the availability on the payers account of the amount necessary to execute the payment transaction, as defined in Art. 65 PSD2.",
        "operationId" : "getConfirmationOfFunds",
        "parameters" : [ {
          "$ref" : "#/parameters/acceptEncodingParam"
        }, {
          "$ref" : "#/parameters/acceptLanguageParam"
        }, {
          "$ref" : "#/parameters/acceptCharsetParam"
        }, {
          "$ref" : "#/parameters/xjwsSignatureParam"
        }, {
          "$ref" : "#/parameters/xRequestIdParam"
        }, {
          "name" : "confirmationOfFundsRequest",
          "in" : "body",
          "description" : "Object with amount to check",
          "required" : true,
          "schema" : {
            "$ref" : "#/definitions/ConfirmationOfFundsRequest"
          }
        } ],
        "tags" : [ "CAF" ],
        "responses" : {
          "200" : {
            "description" : "Success",
            "schema" : {
              "$ref" : "#/definitions/ConfirmationOfFundsResponse"
            },
            "headers" : {
              "Content-Encoding" : {
                "type" : "string",
                "description" : "Gzip, deflate"
              },
              "X-JWS-SIGNATURE" : {
                "type" : "string",
                "description" : "Detached JWS signature of the body of the response"
              }
            }
          },
          "400" : {
            "description" : "Bad request",
            "schema" : {
              "$ref" : "#/definitions/Error"
            }
          },
          "403" : {
            "description" : "Forbidden",
            "schema" : {
              "$ref" : "#/definitions/Error"
            }
          },
          "405" : {
            "description" : "Method Not Allowed",
            "schema" : {
              "$ref" : "#/definitions/Error"
            }
          },
          "406" : {
            "description" : "Not Acceptable",
            "schema" : {
              "$ref" : "#/definitions/Error"
            }
          },
          "415" : {
            "description" : "Unsupported Media Type",
            "schema" : {
              "$ref" : "#/definitions/Error"
            }
          },
          "422" : {
            "description" : "Unprocessable entity",
            "schema" : {
              "$ref" : "#/definitions/Error"
            }
          },
          "429" : {
            "description" : "Request limit for the requested service exceeded",
            "schema" : {
              "$ref" : "#/definitions/Error"
            }
          },
          "500" : {
            "description" : "Internal Server Error",
            "schema" : {
              "$ref" : "#/definitions/Error"
            }
          },
          "501" : {
            "description" : "Not Implemented",
            "schema" : {
              "$ref" : "#/definitions/Error"
            }
          },
          "503" : {
            "description" : "Service Unavailable",
            "schema" : {
              "$ref" : "#/definitions/Error"
            }
          }
        }
      }
    }
  },
  "securityDefinitions" : {
    "xs2a_auth_aspsp" : {
      "type" : "oauth2",
      "flow" : "accessCode",
      "authorizationUrl" : "https://apiHost/v2_1_2.1/auth/v2_1_2.1/authorize",
      "tokenUrl" : "https://apiHost/v2_1_2.1/auth/v2_1_2.1/token",
      "scopes" : {
        "ais-accounts" : "Permission to retrieve the list of accounts through Account Information Service",
        "ais" : "Permission to execute methods of Account Information Service in context of particular account",
        "pis" : "Permission to initiate payments and check their statuses through Payment Initiation Service"
      }
    },
    "xs2a_auth_decoupled" : {
      "type" : "oauth2",
      "flow" : "accessCode",
      "authorizationUrl" : "https://apiHost/auth/authorizeExt",
      "tokenUrl" : "https://apiHost/v2_1_2.1/auth/v2_1_2.1/token",
      "scopes" : {
        "ais-accounts" : "Permission to retrieve the list of accounts through Account Information Service",
        "ais" : "Permission to execute methods of Account Information Service in context of particular account",
        "pis" : "Permission to initiate payments and check their statuses through Payment Initiation Service"
      }
    }
  },
  "definitions" : {
...
    "ConfirmationOfFundsRequest" : {
      "description" : "Klasa zapytania o dostępne środki płatnicze na rachunku / Confirmation of Funds Request Class",
      "type" : "object",
      "required" : [ "requestHeader", "accountNumber", "amount", "currency" ],
      "properties" : {
        "requestHeader" : {
          "$ref" : "#/definitions/RequestHeaderWithoutToken"
        },
        "accountNumber" : {
          "type" : "string",
          "maxLength" : 34,
          "description" : "Numer konta / Account number"
        },
        "amount" : {
          "type" : "string",
          "pattern" : "^(0|([1-9][0-9]{0,}))\\.\\d{2}$",
          "description" : "Wielkość środków której dotyczy zaptanie / Amount of the transaction"
        },
        "currency" : {
          "type" : "string",
          "maxLength" : 3,
          "description" : "Kod ISO Waluty (waluta transakcji) / Currency of transaction (ISO)"
        }
      },
      "xml" : {
        "name" : "ConfirmationOfFundsRequest"
      }
    },
    "ConfirmationOfFundsResponse" : {
      "description" : "Klasa odpowiedzi na zapytanie o dostępne środki płatnicze na rachunku / Confirmation of Funds Response Class",
      "required" : [ "responseHeader", "fundsAvailable" ],
      "properties" : {
        "responseHeader" : {
          "$ref" : "#/definitions/ResponseHeader"
        },
        "fundsAvailable" : {
          "description" : "Status - Czy środki są dostępne / Status - are funds available",
          "type" : "boolean",
          "enum" : [ true, false ]
        }
      }
    },
    "Error" : {
      "description" : "Klasa zawierająca informacje o zwracanym błędzie / Error class",
      "properties" : {
        "responseHeader" : {
          "$ref" : "#/definitions/ResponseHeader"
        },
        "code" : {
          "type" : "string",
          "description" : "Kod błędu / Error code"
        },
        "message" : {
          "type" : "string",
          "description" : "Opis błędu / Error message"
        }
      }
    },
...
    "RequestHeaderWithoutToken" : {
      "description" : "Klasa zawierająca informacje o PSU / PSU Information Class",
      "type" : "object",
      "required" : [ "requestId" ],
      "properties" : {
        "requestId" : {
          "type" : "string",
          "format" : "uuid",
          "description" : "Identyfikator żądania w formacie UUID (Wariant 1, Wersja 1) zgodnym ze standardem RFC 4122 / Request ID using UUID format (Variant 1, Version 1) described in RFC 4122 standard"
        },
        "userAgent" : {
          "type" : "string",
          "description" : "Browser agent dla PSU / PSU browser agent"
        },
        "ipAddress" : {
          "type" : "string",
          "description" : "Adres IP końcowego urządzenia PSU. Wymagany dla isDirectPsu=true. / IP address of PSU's terminal device. Required when isDirectPsu=true."
        },
        "sendDate" : {
          "type" : "string",
          "format" : "date-time",
          "description" : "Oryginalna data wysłania, format: YYYY-MM-DDThh:mm:ss[.mmm] / Send date"
        },
        "tppId" : {
          "type" : "string",
          "description" : "Identyfikator TPP / TPP ID"
        }
      }
    },
...
    "ResponseHeader" : {
      "description" : "Klasa zawierająca zwrotne metadane / Metadata Class",
      "properties" : {
        "requestId" : {
          "type" : "string",
          "format" : "uuid",
          "description" : "Identyfikator żądania otrzymanego od TPP / Identifier of request received from TPP"
        },
        "sendDate" : {
          "type" : "string",
          "format" : "date-time",
          "description" : "Data odpowiedzi, format: YYYY-MM-DDThh:mm:ss[.mmm] / Send date"
        },
        "isCallback" : {
          "type" : "boolean",
          "description" : "Znacznik określający czy odpowiedz zostanie przekazana w formie wywołania zwrotnego. true - gdy odpowiedz w formie wywołania zwrotnego. Inna dopuszczalna wartość to false."
        }
      }
    },
...
  }
}